Phase 7: 32-bit support (33-bit side + wide residual) — all bit depths done
Carry the channel signal as i64 throughout (mid-side, fixed, lpc, verbatim,
warmup) so the 33-bit side (L-R) and 32-bit mid ((L+R)>>1) never overflow;
<=24-bit values fit i32 and stay byte-identical.

Five libFLAC quirks matched exactly to reach byte-parity:
- Constant detection disabled at subframe_bps >= 28 (the _limit_residual
  predictor reports rbps[1]=34.0, so a constant 32-bit signal is FIXED).
- get_wasted_bits_wide_ returns shift 1 for an all-zero 33-bit side.
- Fixed residuals computed in i64 and wrapped to i32 (no bail); order guess
  uses per-order validity (|residual| > i32::MAX invalidates an order).
- Fixed subframe skipped when its estimated bits/sample >= subframe_bps
  (stream_encoder.c:3561) -> VERBATIM for incompressible wide signals.
- Rice parameter selection short-circuits on mean < 2 (also fixes a latent
  all-zero-partition multiply-overflow panic).

Tests: bit_depths_match_c + wider_depth_full_stream_matches_c now include
32-bit; new wide_32bit_channel_cases_match_c (identical/anti/independent/
near-correlated L-R). 18 unit + 32 differential pass; fmt + clippy clean.

+++ src/lpc/mod.rs
@@ -25,10 +25,13 @@ pub use quantize::Quantized;
 /// the predictor history are sized to this.
 pub const MAX_LPC_ORDER: usize = 32;
 
-/// Apply a window to integer signal data (`FLAC__lpc_window_data`, `lpc.c:68`):
-/// `out[i] = in[i] * window[i]`, the integer promoted to `f32` and multiplied in
-/// `f32`. `in`/`window`/`out` are all at least `data_len` long.
-pub fn window_data(input: &[i32], window: &[f32], out: &mut [f32], data_len: usize) {
+/// Apply a window to integer signal data (`FLAC__lpc_window_data[_wide]`,
+/// `lpc.c:68`): `out[i] = in[i] * window[i]`, the integer promoted to `f32` and
+/// multiplied in `f32`. `in` is `i64` (the unified channel type — the 33-bit side
+/// loses precision in the `f32` cast exactly as the C's `_wide` variant does; for
+/// ≤24-bit values the cast is identical to the narrow path). `in`/`window`/`out`
+/// are all at least `data_len` long.
+pub fn window_data(input: &[i64], window: &[f32], out: &mut [f32], data_len: usize) {
     for i in 0..data_len {
         out[i] = input[i] as f32 * window[i];
     }
@@ -41,7 +44,7 @@ pub fn window_data(input: &[i32], window: &[f32], out: &mut [f32], data_len: usi
 /// trailing zero; only `out[0..2*part_size)` is subsequently read by the
 /// autocorrelation, but the exact indexing is preserved.
 pub fn window_data_partial(
-    input: &[i32],
+    input: &[i64],
     window: &[f32],
     out: &mut [f32],
     data_len: usize,
